pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "myhttpserver"
version = "0.0.1"
description = ""
dependencies = [
    "aiofiles>=25.1.0",
    "httpx>=0.28.1",
    "pyhcl>=0.4.5",
    "uvloop>=0.22.1",
    "xxhash>=3.6.0",
    "zstandard>=0.25.0",
]

[project.scripts]
myhttpserver = "src.main:main"

[tool.setuptools.packages.find]
include = ["src", "src.*"]

[tool.pytest.ini_options]
pythonpath = ["."]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "ipdb>=0.13.13",
    "py-spy>=0.4.1",
    "pytest>=8.0.0",
    "requests>=2.32.5",
    "rich>=14.3.2",
    "ruff>=0.15.1",
    "snakeviz>=2.2.2",
    "watchfiles>=1.1.1",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
